topic: code enhanced type: functionality module: commons description: Added stroke rendering of shapes out of svg or Mark elements.

+++ src/org/deegree/rendering/r2d/strokes/ShapeStroke.java
@@ -33,19 +33,19 @@
  e-mail: info@deegree.org
 ----------------------------------------------------------------------------*/
 /*
-Copyright 2006 Jerry Huxtable
+ Copyright 2006 Jerry Huxtable
 
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
+ Licensed under the Apache License, Version 2.0 (the "License");
+ you may not use this file except in compliance with the License.
+ You may obtain a copy of the License at
 
-   http://www.apache.org/licenses/LICENSE-2.0
+ http://www.apache.org/licenses/LICENSE-2.0
 
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
+ Unless required by applicable law or agreed to in writing, software
+ distributed under the License is distributed on an "AS IS" BASIS,
+ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ See the License for the specific language governing permissions and
+ limitations under the License.
  */
 
 package org.deegree.rendering.r2d.strokes;
@@ -60,17 +60,17 @@ import java.awt.geom.Rectangle2D;
 
 /**
  * <code>ShapeStroke</code>
- *
+ * 
  * @author Jerry Huxtable
  * @author <a href="mailto:schmitz@lat-lon.de">Andreas Schmitz</a>
  * @author last edited by: $Author$
- *
+ * 
  * @version $Revision$, $Date$
  */
 public class ShapeStroke implements Stroke {
     private Shape shapes[];
 
-    private float advance;
+    private double advance;
 
     private boolean repeat = true;
 
@@ -82,7 +82,7 @@ public class ShapeStroke implements Stroke {
      * @param shapes
      * @param advance
      */
-    public ShapeStroke( Shape shapes, float advance ) {
+    public ShapeStroke( Shape shapes, double advance ) {
         this( new Shape[] { shapes }, advance );
     }
 
@@ -90,7 +90,7 @@ public class ShapeStroke implements Stroke {
      * @param shapes
      * @param advance
      */
-    public ShapeStroke( Shape shapes[], float advance ) {
+    public ShapeStroke( Shape shapes[], double advance ) {
         this.advance = advance;
         this.shapes = new Shape[shapes.length];
